+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.2] - 2026-02-21
+### Added
+- Optimize (~twice faster) Levenshtein and Damerau-Levenshtein using numba library through njit.
+- Distance functions are in a new file: distances.py.
+
 ## [0.4.1] - 2026-02-19
 ### Added
 - Progression bar for both database compilation and authors comparison.